+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# minipdf 0.2.2
+
+* draw state parameters are now set via their own methods e.g. `obj$fill('black')`, 
+  however using `update()` still works e.g. `obj$update(fill = 'black')`
+
 # minipdf 0.2.1
 
 * transformations are now methods directly on the PDFStream object rather than attributes